Signifcantly improved tests
Also moved date formatting code from the parser_* class to calendar.py,
which makes it easier to consistently compare dates.

Expanded tests to check more of the returned data.  Tests can now
include a file with expected results (serialized as JSON) which makes  it
simpler to compare lots of events.